removed rapydo core (#9)
* Bump: 0.6.0

* [rapydo/issues#76] Fix to use templates with new naming convention

* Tests executed in development mode

* Code cleanup

* Moving test utilities from old nose2 class to new pytest clsss

* Utility to load group of variables not related to services

* Allow custom init in our complex definition

* Adding the rabbit queue to services

* Moving test utilities from nose2 to pytest

* Developing new tests [note: this test will fail]

* Utility for basic tests on endpoint

* New test for html responses, testing admin users endpoitn

* BUG fix, unable to perform json decode on html output

* Bug fix tests

* [rapydo/issues#51] Testing HTML responses

* Bug fix html tests

* Bug fix tests

* Added some check on swagger spec
Removed back-compatibility fix for choices expansion in getJsonResponse

* Bug fix tests

* Removed string format in logging

* [rapydo/issues#48] Very basic test on admin/users endpoint

* [rapydo/issues#48] Admin users is implemented only for neo4j

* Bug fix

* Fix to skip secondary tests on NOT ALLOWED methods

* [rapydo/issues#48] Admin users is implemented only for neo4j

* Bug fix import

* Bump: 0.6.0

* BUG fix admin users tests

* Renamed utility function, cannot start with "test_"

* Moved an obsolete file to be excluded from coverage, to be deleted in a near future

* Moving test utilities, to be completed

* Bug fix

* Allowed _test_endpoint function to skip tests on specific methods

* Better raise of errors during custom initialization

* Variables: easier passing around

* Testing rabbit with merret

* Function deprecation

* Extending building data to email format

* Fixed a couple of errors from landscape

* Reducing function complexity

* Using %s instead of {}.format

* Removed unused import

* Handling some unused argument

* Trying a bug fix

* Bug fix to check for relationships in responses

* Tests to handle too large range requests

* Minor fix, changed default prefix for random data

* [rapydo/issues#84] Login is now case insensitive

* A test to understand why tests are failing

* [rapydo/issues#84] Fixed tests for case insensitve login

* [rapydo/issues#84] Fixed token deletion

* Bug fix token invalidation on sql

* [rapydo/issues#79] Fixed do_config function to use read.configuration from utilities

* Bug fix invalidate tokens

* [rapydo/issues#79] Renamed rapydo-core as rapydo-tests

* [rapydo/issues#79] Fix clone output

* Bug fix on bool "enable" detection

* More debugging

* change url

* Use the zone method to also append suffix

* [EUDAT-B2STAGE/http-api#109] Update to the latest irods release

* Added post data to basic tests

* Swagger model for admin users

* Better group management in admin users API

* Typo

* Bug fix for admin users authentication

* Allowing group admins to create/modify/delete users [ONLY FOR NEO4J]

* Bug fix check group admin for base model [without coordinator link]

* Disabled user creation for non admins, to be done: check if non admin user is allowed to assign users to che selected group/role

+++ restapi/resources/admin_users.py
@@ -3,13 +3,14 @@
 from restapi import decorators as decorate
 from restapi.services.neo4j.graph_endpoints import GraphBaseOperations
 from restapi.exceptions import RestApiException
-from restapi.services.neo4j.graph_endpoints import graph_transactions
-from restapi.services.neo4j.graph_endpoints import catch_graph_exceptions
+# from restapi.services.neo4j.graph_endpoints import graph_transactions
+# from restapi.services.neo4j.graph_endpoints import catch_graph_exceptions
 from restapi.services.authentication import BaseAuthentication
+from restapi.services.detect import detector
 from utilities import htmlcodes as hcodes
 
 from utilities.logs import get_logger
-logger = get_logger(__name__)
+log = get_logger(__name__)
 
 __author__ = "Mattia D'Antonio (m.dantonio@cineca.it)"
 
@@ -19,7 +20,7 @@ class AdminUsers(GraphBaseOperations):
     def link_role(self, user, properties):
         ids = self.parseAutocomplete(
             properties, 'roles', id_key='name', split_char=',')
-        logger.critical(ids)
+        # log.critical(ids)
 
         if ids is None:
             return
@@ -34,37 +35,7 @@ class AdminUsers(GraphBaseOperations):
             except self.graph.Role.DoesNotExist:
                 pass
 
-    @decorate.catch_error()
-    @catch_graph_exceptions
-    def get(self, id=None):
-
-        self.initGraph()
-        nodeset = self.graph.User.nodes
-
-        data = []
-        for n in nodeset.all():
-            user = self.getJsonResponse(n, max_relationship_depth=2)
-            data.append(user)
-
-        return self.force_response(data)
-
-    @decorate.catch_error()
-    @catch_graph_exceptions
-    @graph_transactions
-    def post(self):
-
-        self.initGraph()
-
-        v = self.get_input()
-        if len(v) == 0:
-            raise RestApiException(
-                'Empty input',
-                status_code=hcodes.HTTP_BAD_REQUEST)
-
-        schema = self.get_endpoint_custom_definition()
-        # INIT #
-        properties = self.read_properties(schema, v)
-
+    def parse_group(self, v):
         groups = self.parseAutocomplete(v, 'group', id_key='id')
 
         if groups is None:
@@ -72,30 +43,133 @@ class AdminUsers(GraphBaseOperations):
                 'Group not found', status_code=hcodes.HTTP_BAD_REQUEST)
 
         group_id = groups.pop()
-
         group = self.graph.Group.nodes.get_or_none(uuid=group_id)
-        # group = self.getNode(self.graph.Group, group_id, field='uuid')
 
         if group is None:
             raise RestApiException(
                 'Group not found', status_code=hcodes.HTTP_BAD_REQUEST)
 
+        return group
+
+    def check_permissions(self, user, node, is_admin, is_group_admin):
+
+        if node is None:
+            return False
+
+        # an ADMIN is always authorized
+        if is_admin:
+            return True
+
+        # You are neither an ADMIN nor a GROUP ADMIN
+        if not is_group_admin:
+            return False
+
+        # If you are not an ADMIN, you cannot modify yourself...
+        # use the profile instead!
+        if user == node:
+            return False
+
+        # FIXME: only implemented for neo4j
+        # You are a group admin... but the group mathes??
+        for g in user.coordinator.all():
+            if node.belongs_to.is_connected(g):
+                return True
+
+        return False
+
+    @decorate.catch_error()
+    # @catch_graph_exceptions
+    def get(self, id=None):
+
+        data = []
+        if not detector.check_availability('neo4j'):
+            log.warning("This endpoint is implemented only for neo4j")
+            return self.force_response(data)
+
+        self.graph = self.get_service_instance('neo4j')
+
+        is_admin = self.auth.verify_admin()
+        is_group_admin = self.auth.verify_group_admin()
+        if not is_admin and not is_group_admin:
+            raise RestApiException(
+                "You are not authorized: missing privileges",
+                status_code=hcodes.HTTP_BAD_UNAUTHORIZED)
+
+        current_user = self.get_current_user()
+        nodeset = self.graph.User.nodes
+
+        for n in nodeset.all():
+
+            is_authorized = self.check_permissions(
+                current_user, n, is_admin, is_group_admin
+            )
+            if not is_authorized:
+                continue
+
+            user = self.getJsonResponse(n, max_relationship_depth=2)
+            data.append(user)
+
+        return self.force_response(data)
+
+    @decorate.catch_error()
+    # @catch_graph_exceptions
+    # @graph_transactions
+    def post(self):
+
+        v = self.get_input()
+        if len(v) == 0:
+            raise RestApiException(
+                'Empty input',
+                status_code=hcodes.HTTP_BAD_REQUEST)
+
+        if not detector.check_availability('neo4j'):
+            log.warning("This endpoint is implemented only for neo4j")
+            return self.force_response('0')
+
+        self.graph = self.get_service_instance('neo4j')
+
+        is_admin = self.auth.verify_admin()
+        is_group_admin = self.auth.verify_group_admin()
+        if not is_admin and not is_group_admin:
+            raise RestApiException(
+                "You are not authorized: missing privileges",
+                status_code=hcodes.HTTP_BAD_UNAUTHORIZED)
+
+        schema = self.get_endpoint_custom_definition()
+        # INIT #
+        properties = self.read_properties(schema, v)
+
+        group = None
+        if 'group' in v:
+            group = self.parse_group(v)
+
         # GRAPH #
         properties["authmethod"] = "credentials"
-        # if "password" in properties:
-        properties["password"] = \
-            BaseAuthentication.hash_password(properties["password"])
+        if "password" in properties:
+            properties["password"] = \
+                BaseAuthentication.hash_password(properties["password"])
         # properties["name_surname"] = \
-        #     self.createUniqueIndex(properties["name"], properties["surname"])
+        #     self.createUniqueIndex(
+        #         properties["name"], properties["surname"])
         user = self.graph.User(**properties).save()
-        user.belongs_to.connect(group)
+
+        if group is not None:
+            if not is_admin:
+                raise RestApiException(
+                    "Check if you are allowed to assign users to this group")
+
+            user.belongs_to.connect(group)
+
+        if not is_admin:
+            raise RestApiException(
+                "Check if you are allowed to assign users to this role")
         self.link_role(user, v)
 
         return self.force_response(user.uuid)
 
     @decorate.catch_error()
-    @catch_graph_exceptions
-    @graph_transactions
+    # @catch_graph_exceptions
+    # @graph_transactions
     def put(self, user_id=None):
 
         if user_id is None:
@@ -104,15 +178,35 @@ class AdminUsers(GraphBaseOperations):
                 "Please specify a user id",
                 status_code=hcodes.HTTP_BAD_REQUEST)
 
+        if not detector.check_availability('neo4j'):
+            log.warning("This endpoint is implemented only for neo4j")
+            return self.empty_response()
+
         schema = self.get_endpoint_custom_definition()
-        self.initGraph()
+        self.graph = self.get_service_instance('neo4j')
+
+        is_admin = self.auth.verify_admin()
+        is_group_admin = self.auth.verify_group_admin()
+        if not is_admin and not is_group_admin:
+            raise RestApiException(
+                "You are not authorized: missing privileges",
+                status_code=hcodes.HTTP_BAD_UNAUTHORIZED)
 
         v = self.get_input()
 
         user = self.graph.User.nodes.get_or_none(uuid=user_id)
         # user = self.getNode(self.graph.User, user_id, field='uuid')
         if user is None:
-            raise RestApiException("User not found")
+            raise RestApiException(
+                "This user cannot be found or you are not authorized")
+
+        current_user = self.get_current_user()
+        is_authorized = self.check_permissions(
+            current_user, user, is_admin, is_group_admin
+        )
+        if not is_authorized:
+            raise RestApiException(
+                "This user cannot be found or you are not authorized")
 
         if "password" in v and v["password"] == "":
             del v["password"]
@@ -122,13 +216,14 @@ class AdminUsers(GraphBaseOperations):
         self.update_properties(user, schema, v)
         user.name_surname = self.createUniqueIndex(user.name, user.surname)
         user.save()
-        groups = self.parseAutocomplete(v, 'group', id_key='id')
 
-        if groups is not None:
-            group_id = groups.pop()
+        if 'group' in v:
 
-            group = self.graph.Group.nodes.get_or_none(uuid=group_id)
-            # group = self.getNode(self.graph.Group, group_id, field='uuid')
+            group = self.parse_group(v)
+
+            if not is_admin:
+                raise RestApiException(
+                    "Check if you are allowed to assign users to this group")
 
             p = None
             for p in user.belongs_to.all():
@@ -139,13 +234,18 @@ class AdminUsers(GraphBaseOperations):
                 user.belongs_to.reconnect(p, group)
             else:
                 user.belongs_to.connect(group)
+
+        if not is_admin:
+            raise RestApiException(
+                "Check if you are allowed to assign users to this role")
+
         self.link_role(user, v)
 
         return self.empty_response()
 
     @decorate.catch_error()
-    @catch_graph_exceptions
-    @graph_transactions
+    # @catch_graph_exceptions
+    # @graph_transactions
     def delete(self, user_id=None):
 
         if user_id is None:
@@ -154,40 +254,33 @@ class AdminUsers(GraphBaseOperations):
                 "Please specify a user id",
                 status_code=hcodes.HTTP_BAD_REQUEST)
 
-        self.initGraph()
+        if not detector.check_availability('neo4j'):
+            log.warning("This endpoint is implemented only for neo4j")
+            return self.empty_response()
+
+        self.graph = self.get_service_instance('neo4j')
+
+        is_admin = self.auth.verify_admin()
+        is_group_admin = self.auth.verify_group_admin()
+        if not is_admin and not is_group_admin:
+            raise RestApiException(
+                "You are not authorized: missing privileges",
+                status_code=hcodes.HTTP_BAD_UNAUTHORIZED)
 
         user = self.graph.User.nodes.get_or_none(uuid=user_id)
         # user = self.getNode(self.graph.User, user_id, field='uuid')
         if user is None:
-            raise RestApiException('This user cannot be found')
+            raise RestApiException(
+                "This user cannot be found or you are not authorized")
+
+        current_user = self.get_current_user()
+        is_authorized = self.check_permissions(
+            current_user, user, is_admin, is_group_admin
+        )
+        if not is_authorized:
+            raise RestApiException(
+                "This user cannot be found or you are not authorized")
 
         user.delete()
 
         return self.empty_response()
-
-
-class UserRole(GraphBaseOperations):
-    @decorate.catch_error(exception=Exception, catch_generic=True)
-    @catch_graph_exceptions
-    def get(self, query=None):
-
-        self.initGraph()
-
-        data = []
-        cypher = "MATCH (r:Role)"
-
-        if query is not None:
-            cypher += " WHERE r.description <> 'automatic'"
-            cypher += " AND r.name =~ '(?i).*%s.*'" % query
-
-        cypher += " RETURN r ORDER BY r.name ASC"
-
-        if query is None:
-            cypher += " LIMIT 20"
-
-        result = self.graph.cypher(cypher)
-        for row in result:
-            r = self.graph.Role.inflate(row[0])
-            data.append({"name": r.name, "description": r.description})
-
-        return self.force_response(data)
